requirements.txt
torch>=1.9.0
pytorch-lightning>=1.5.0
transformers>=4.10.0
datasets>=1.10.0
tokenizers>=0.10.0
tensorboard>=2.5.0
pyyaml>=5.4.0